Changes to make types actually correspond
git-svn-id: https://activeobjects.dev.java.net/svn/activeobjects/trunk@974 69e892e8-b12f-0410-b7e1-b50061bc6545

+++ ActiveObjects/test/net/java/ao/types/DatabaseTypeTest.java
@@ -153,13 +153,13 @@ public class DatabaseTypeTest extends DataTest {
 	@Test
 	public void testDefaultParseValue() throws MalformedURLException {
 		assertEquals(123, new IntegerType().defaultParseValue("123"));
-		assertEquals(123.456, new DoubleType().defaultParseValue("123.456"));
-		assertEquals(123.456, new FloatType().defaultParseValue("123.456"));
+		assertEquals(123.456d, new DoubleType().defaultParseValue("123.456"));
+		assertEquals(123.456f, new FloatType().defaultParseValue("123.456"));
 		assertEquals("My test value", new VarcharType().defaultParseValue("My test value"));
 		assertEquals(new URL("http://www.google.com"), new URLType().defaultParseValue("http://www.google.com"));
 		assertEquals(false, new BooleanType().defaultParseValue("false"));
 		assertEquals(String.class, new ClassType().defaultParseValue("java.lang.String"));
-		assertEquals(123, new TinyIntType().defaultParseValue("123"));
+		assertEquals((short) 123, new TinyIntType().defaultParseValue("123"));
 		assertEquals('c', new CharType().defaultParseValue("c"));
